requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.1.0
Django==5.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0